Added comments to DAL layer and added some extra methods

+++ ChapeauModel/MenuCategory.cs
@@ -10,9 +10,9 @@ namespace ChapeauModel
     {
         public string Id { get; set; }
         public string Name { get; set; }
-        public int VAT { get; set; }
+        public decimal VAT { get; set; }
 
-        public MenuCategory(string id, string name, int vat)
+        public MenuCategory(string id, string name, decimal vat)
         {
             Id = id;
             Name = name;
